rename query explain prompt
